Rename `backpack.utils.einsum` → `backpack.utils.ein`
