+++ backpack/core/derivatives/crossentropyloss.py
@@ -5,7 +5,7 @@ from torch import sqrt as torchsqrt
 from torch.nn import CrossEntropyLoss
 from torch.nn.functional import one_hot
 
-from backpack.utils.einsum import einsum
+from backpack.utils.ein import einsum
 from backpack.core.derivatives.basederivatives import BaseLossDerivatives
