+++ backpack/core/derivatives/elementwise.py
@@ -1,4 +1,4 @@
-from backpack.utils.einsum import einsum
+from backpack.utils.ein import einsum
 from backpack.core.derivatives.basederivatives import BaseDerivatives
