+++ backpack/extensions/firstorder/sum_grad_squared/conv2d.py
@@ -1,6 +1,6 @@
 from backpack.extensions.firstorder.base import FirstOrderModuleExtension
 from backpack.utils import conv as convUtils
-from backpack.utils.einsum import einsum
+from backpack.utils.ein import einsum
 
 
 class SGSConv2d(FirstOrderModuleExtension):
