+++ backpack/extensions/secondorder/diag_hessian/diag_h_base.py
@@ -1,7 +1,7 @@
 from backpack.extensions.mat_to_mat_jac_base import MatToJacMat
 from torch import clamp, diag_embed
 from numpy import prod
-from backpack.utils.einsum import einsum
+from backpack.utils.ein import einsum
 
 
 class DiagHBaseModule(MatToJacMat):
