+++ backpack/extensions/secondorder/hbp/conv2d.py
@@ -1,6 +1,6 @@
 from backpack.core.derivatives.conv2d import Conv2DDerivatives
 from backpack.utils import conv as convUtils
-from backpack.utils.einsum import einsum
+from backpack.utils.ein import einsum
 
 from backpack.extensions.secondorder.hbp.hbpbase import HBPBaseModule
 from backpack.extensions.secondorder.hbp.hbp_options import (
